requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
dnspython==2.0.0
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.21.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.2
python-dotenv==0.15.0
six==1.15.0
Werkzeug==1.0.1
